Chage Polynom interface and simplify one part

+++ lab1/tests.cpp
@@ -4,7 +4,7 @@ extern "C" int64_t Sum(int32_t x, uint8_t y);
 extern "C" bool CheckOverflow(int64_t x, int64_t y);
 extern "C" int64_t ComputeFn(int64_t x, int64_t y);
 extern "C" int64_t Clock(int32_t h, int32_t m, int32_t f);
-extern "C" uint64_t Polynom(int64_t x);
+extern "C" uint64_t Polynom(int32_t x);
 
 // ---------------------------------------------------------
